requirements_automation.txt
# UI自动化和图像分析依赖

# 核心库（虚拟摄像头已有）
opencv-python>=4.8.0
numpy>=1.24.0

# 图像处理
Pillow>=10.0.0

# UI自动化
pyautogui>=0.9.54

# OCR支持（可选）
# pytesseract>=0.3.10

# 屏幕截图
# mss>=9.0.1

